Moved receiver into separate class, moved realtive to reciver method from mainActivity to MusicIntentReceiver class.
